backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP Client
httpx==0.27.2
requests==2.32.3

# CORS & Security
# FastAPI has built-in CORS support

# Environment
python-dotenv==1.0.1

# Utils
python-dateutil==2.9.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.4
